Extract D8 map validation internals
